avcodec/aacenc: Treat single channel as mono
A single channel is normally interpreted as mono.
However if channelsplit filter is used, the individual
channels are given their channel mask (ex: FL) as channel layout.
This can cause issue with the native encoder because only channel
layouts in spec will be encoded without PCE.
(they are listed in aac_normal_chan_layouts[]).
To avoid that, the layout for a single channel is set to mono and
therefore its encoding is achieved without PCE.
This fixes ticket 7266.